pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "Macrostrat Python libraries"
name = "macrostrat.python_libraries"
version = "1.3.1"
packages = [
  {include = "macrostrat"},
]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
"macrostrat.app_frame" = {path = "./app-frame", develop = true}
"macrostrat.database" = {path = "./database", develop = true}
"macrostrat.dinosaur" = {path = "./dinosaur", develop = true}
"macrostrat.package_tools" = {path = "./package-tools", develop = true}
"macrostrat.utils" = {path = "./utils", develop = true}
pytest = "^7.2.2"
python-dotenv = "^1.0.0"
requests = "^2.27.1"
rich = "^13"
toml = "^0.10.2"
isort = "^5.13.2"

[tool.pytest.ini_options]
addopts = "--confcutdir=."

[tool.poetry.scripts]
test-app = "test_app:main"
format = "isort -rc . && black ."

[tool.isort]
profile = "black"
known_first_party = "macrostrat"